+++ bundles/framework/layerlist/resources/locale/en.js
@@ -63,20 +63,6 @@ Oskari.registerLocalization(
                 'UNKNOWN': 'The status information for this map layer is not available.',
                 'UNSTABLE': ''
             },
-            'layerTooltipTitle': {
-                'wms': 'Raster layer',
-                'wmts': 'Raster layer',
-                'arcgis93': 'Raster layer',
-                'arcgis': 'Raster layer',
-                'vectortile': 'Raster layer',
-                'bingmaps': 'Raster layer',
-                'wfs': 'Vector layer',
-                'vector': 'Vector layer',
-                'userlayer': 'Own dataset',
-                'myplaces': 'My map layer',
-                'analysislayer': 'Own analysis',
-                'tiles3d': '3D layer'
-            },
             'guidedTour': {
                 'title': 'Map layers',
                 'message': 'In the Map Layers menu you can find all the map layers in the map service.<br/><br/>Sort map layers by theme or data provider.<br/><br/>Search map layers by a map layer name, a data provider name or a keyword. You can browse the newest map layers or filter map layers to find what you need easier.<br/><br/>You can check currently open map layers in the Selected Layers tab. You can rearrange the layers by dragging, adjust their opacity and change the styling of some layers.',
