+++ bundles/framework/layerlist/resources/locale/sv.js
@@ -62,20 +62,6 @@ Oskari.registerLocalization(
                 'UNKNOWN': 'Statusinformationen för det här kartlagret är inte tillgängligt.',
                 'UNSTABLE': ''
             },
-            'layerTooltipTitle': {
-                'wms': 'Kartlager i rasterformat',
-                'wmts': 'Kartlager i rasterformat',
-                'arcgis93': 'Kartlager i rasterformat',
-                'arcgis': 'Kartlager i rasterformat',
-                'vectortile': 'Kartlager i rasterformat',
-                'bingmaps': 'Kartlager i rasterformat',
-                'wfs': 'Kartlager i vektorformat',
-                'vector': 'Kartlager i vektorformat',
-                'userlayer': 'Eget dataset',
-                'myplaces': 'Mitt kartlager',
-                'analysislayer': 'Min analys',
-                'tiles3d': '3D kartlager'
-            },
             'guidedTour': {
                 'title': 'Kartlager',
                 'message': 'På Kartlager-listan hittar du alla i karttjänsten tillgängliga kartlager.<br/><br/>Sortera kartlager antingen med tema eller med dataproducent.<br/><br/>Sök kartlager med kartlagernamn, dataproducents namn eller sökord. Du kan bläddra i de nyaste kartlager eller filtrera kartlager för att bättre hitta vad du söker.<br/><br/>Bladet Valda kartlager visar vilka kartlager är öppet. Du kan dra dessa kartlager i en annan ordning, justera kartlagrens opacitet eller byta stil på en del kartlager.',
