+++ bundles/framework/oskariui/resources/locale/en.js
@@ -136,6 +136,20 @@ Oskari.registerLocalization({
                 lessThan: 'less than',
                 atMost: 'at most'
             }
+        },
+        layerTooltipTitle: {
+            'wms': 'Raster layer',
+            'wmts': 'Raster layer',
+            'arcgis93': 'Raster layer',
+            'arcgis': 'Raster layer',
+            'vectortile': 'Raster layer',
+            'bingmaps': 'Raster layer',
+            'wfs': 'Vector layer',
+            'vector': 'Vector layer',
+            'userlayer': 'Own dataset',
+            'myplaces': 'My map layer',
+            'analysislayer': 'Own analysis',
+            'tiles3d': '3D layer'
         }
     }
 });
