+++ bundles/framework/oskariui/resources/locale/fi.js
@@ -136,6 +136,20 @@ Oskari.registerLocalization({
                 lessThan: 'pienempi kuin',
                 atMost: 'enintään'
             }
+        },
+        layerTooltipTitle: {
+            'wms': 'Rasteritaso',
+            'wmts': 'Rasteritaso',
+            'arcgis93': 'Rasteritaso',
+            'arcgis': 'Rasteritaso',
+            'vectortile': 'Rasteritaso',
+            'bingmaps': 'Rasteritaso',
+            'wfs': 'Vektoritaso',
+            'vector': 'Vektoritaso',
+            'userlayer': 'Oma aineisto',
+            'myplaces': 'Oma karttataso',
+            'analysislayer': 'Oma analyysi',
+            'tiles3d': '3D-taso'
         }
     }
 });
