+++ bundles/framework/oskariui/resources/locale/sv.js
@@ -136,6 +136,20 @@ Oskari.registerLocalization({
                 lessThan: 'mindre än',
                 atMost: 'som mest'
             }
+        },
+        layerTooltipTitle: {
+            'wms': 'Kartlager i rasterformat',
+            'wmts': 'Kartlager i rasterformat',
+            'arcgis93': 'Kartlager i rasterformat',
+            'arcgis': 'Kartlager i rasterformat',
+            'vectortile': 'Kartlager i rasterformat',
+            'bingmaps': 'Kartlager i rasterformat',
+            'wfs': 'Kartlager i vektorformat',
+            'vector': 'Kartlager i vektorformat',
+            'userlayer': 'Eget dataset',
+            'myplaces': 'Mitt kartlager',
+            'analysislayer': 'Min analys',
+            'tiles3d': '3D kartlager'
         }
     }
 });
