+++ src/react/components/icons/EyeOpen.jsx
@@ -1,0 +1,5 @@
+import { EyeFilled } from '@ant-design/icons';
+
+export const EyeOpen = (props) => (
+    <EyeFilled style={{ fontSize: '18px', color: '#006ce8', marginRight: '5px', marginTop: '2px' }} {...props} />
+);
