+++ src/react/components/icons/EyeShut.jsx
@@ -1,0 +1,5 @@
+import { EyeInvisibleOutlined } from '@ant-design/icons';
+
+export const EyeShut = (props) => (
+    <EyeInvisibleOutlined style={{ fontSize: '18px', marginRight: '5px', marginTop: '2px' }} {...props} />
+);
