+++ src/react/components/icons/index.js
@@ -10,5 +10,5 @@ export { MeasureArea as MeasureAreaIcon } from './MeasureArea';
 export { MeasureLine as MeasureLineIcon } from './MeasureLine';
 export { Drag as DragIcon } from './Drag';
 export { LayerIcon } from './LayerIcon';
-export { EyeOpen } from './LayerIcons';
-export { EyeShut } from './LayerIcons';
+export { EyeOpen } from './EyeOpen';
+export { EyeShut } from './EyeShut';
